etl/requirements.txt
redis==5.0.8
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
psycopg[binary]==3.1.18
elasticsearch==8.15.0